refactor api rename endpoint convert post to patch remove service interface
