+++ dream-team-mkt-business/src/main/java/co/simplon/dreamteam/mkt/controllers/DreamTeamIdController.java
@@ -3,7 +3,7 @@ package co.simplon.dreamteam.mkt.controllers;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -28,9 +28,9 @@ public class DreamTeamIdController {
 		return service.get();
 	}
 
-	@PostMapping("/updateData")
+	@PatchMapping("/updateData")
 	@ResponseStatus(code = HttpStatus.OK)
-	public DreamTeamIdDto postMethodName(@RequestBody DreamTeamIdDto inputs) {
+	public DreamTeamIdDto update(@RequestBody DreamTeamIdDto inputs) {
 		return service.updateData(inputs);
 	}
