+++ dream-team-mkt-business/src/main/java/co/simplon/dreamteam/mkt/services/DreamTeamIdService.java
@@ -1,9 +1,25 @@
 package co.simplon.dreamteam.mkt.services;
 
+import org.springframework.stereotype.Service;
+
 import co.simplon.dreamteam.mkt.dtos.DreamTeamIdDto;
+import co.simplon.dreamteam.mkt.mapper.DreamTeamIdMapper;
+import co.simplon.dreamteam.mkt.repositories.DreamTeamIdRepository;
 
-public interface DreamTeamIdService {
-	DreamTeamIdDto get();
+@Service
+public class DreamTeamIdService {
+	private final DreamTeamIdRepository repository;
 
-	DreamTeamIdDto updateData(DreamTeamIdDto dreamTeamIdDto);
+	public DreamTeamIdService(DreamTeamIdRepository repository) {
+		this.repository = repository;
+	}
+
+	public DreamTeamIdDto get() {
+		return DreamTeamIdMapper.entityToDto(repository.findAll().getLast());
+	}
+
+	public DreamTeamIdDto updateData(DreamTeamIdDto dreamTeamIdDto) {
+		return DreamTeamIdMapper.entityToDto(repository.saveAndFlush(DreamTeamIdMapper.updateEntityWithDto(dreamTeamIdDto, repository.findAll().getLast())));
+	}
+
 }
