+++ dream-team-mkt-business/src/main/resources/application-local.properties
@@ -1,0 +1,18 @@
+spring.datasource.url=jdbc:postgresql://localhost:5432/dream_team_mkt
+spring.datasource.username=postgres
+spring.datasource.password=postgres
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+
+#custom proprietie
+co.simplon.dream-team.email.from=no-reply.dev@readresolve.io
+
+# SMTP/Mail properties:
+spring.mail.protocol=smtp
+spring.mail.host=SSL0.OVH.NET
+spring.mail.port=465
+spring.mail.username=no-reply.dev@readresolve.io
+spring.mail.password=Xd9*4LkjE!
+spring.mail.properties.mail.smtp.ssl.enable=true
+
+co.simplon.dream-team.cors="*"
