+++ dream-team-mkt-presentation/src/services/informPanels.js
@@ -12,7 +12,7 @@ class InformPanels {
 
   static async updateData(data) {
     let returnData = {}
-    returnData = await axiosClient.post("/dreamteam/updateData",
+    returnData = await axiosClient.patch("/dreamteam/updateData",
       data
     )
     return returnData.data;
